\name{id}
\encoding{UTF-8}
\Rdversion{1.1}
\docType{methods}
\alias{id,TranscriptExpressionSet-method}
% \alias{id<-,TranscriptExpressionSet-method}
\alias{id}
\title{
   Accessor function for the 'id' slot of a TranscriptExpressionSet object.
}

\description{
   The id slot contains the id for the gene represented in a given TranscriptExpressionSet object.
}

\usage{
\S4method{id}{TranscriptExpressionSet}(object)
% \S4method{id}{TranscriptExpressionSet}(object) <- value
}

\arguments{
  \item{object}{a \code{TranscriptExpressionSet} object.}
}

\author{
   Mar Gonzàlez-Porta, <mar@ebi.ac.uk>
}
\examples{
}   
   
