\name{Ipsum-package}
\encoding{UTF-8}
\alias{Ipsum-package}
\alias{Ipsum}
\docType{package}
\title{
	Ipsum: visualisation of gene and transcript expression levels from RNA-seq experiments
}
\description{
	Plot gene and transcript expression levels estimated from RNA-seq experiments using both sample-specific and summarised approaches.
}
\details{
	\tabular{ll}{
		Package: \tab Ipsum\cr
		Type: \tab Package\cr
		Version: \tab 1.0\cr
		Date: \tab 2014-01-17\cr
		License: \tab GPLv2\cr
	}
}
\author{
	Author: Mar Gonzàlez-Porta <mar@ebi.ac.uk>
}
\references{https://github.com/mgonzalezporta/ipsum}
\keyword{ plot RNA-seq expression splicing }
\examples{}
