\name{readSignificantEvents}
\Rdversion{1.1}
\encoding{UTF-8}
\docType{methods}
\alias{readSignificantEvents}

\title{
   For a given gene, load a list of transcripts that undergo differential usage across conditions.
}
\description{
   Loads a list of transcripts that undergo differential usage across conditions for the specified gene.
}
\usage{
readSignificantEvents(gId=gId, infile=infile)
}
\arguments{
  \item{gId}{Id of the gene for which to load the list of differentially used transcripts.}
  \item{infile}{Path to the file that contains the differential usage data.}
}
\author{
   Mar Gonzàlez-Porta, <mar@ebi.ac.uk>
}
\examples{
}
