\name{texp}
\encoding{UTF-8}
\Rdversion{1.1}
\docType{methods}
\alias{texp,TranscriptExpressionSet-method}
\alias{texp}
\title{
   Accessor function for the 'texp' slot of a TranscriptExpressionSet object.
}

\description{
   The texp slot displays a matrix with the initial expression values provided by the user. Each sample is represented as a column, and each transcript as a row.
}

\usage{
\S4method{texp}{TranscriptExpressionSet}(object)
}

\arguments{
  \item{object}{a \code{TranscriptExpressionSet} object.}
}

\author{
   Mar Gonzàlez-Porta, <mar@ebi.ac.uk>
}
\examples{
}   
   
